--- src/components/RoundFrame.tsx
@@ -1,6 +1,0 @@
-interface Props {
-  children: React.ReactNode;
-}
-export default function RoundFrame({ children }: Props) {
-  return <div>{children}</div>;
-}
