--- src/styles/Home.module.css
@@ -1,6 +1,0 @@
-.header{
-  color:var(--background)
-}
-/* 仮でおいてます.
-消してオッケーです .
-ヘッダーについてるクラスも同様*/
